+++ lib/changes/index.js
@@ -1,0 +1,21 @@
+import insertTable from './insertTable';
+import insertRow from './insertRow';
+import removeRow from './removeRow';
+import insertColumn from './insertColumn';
+import removeColumn from './removeColumn';
+import removeTable from './removeTable';
+import moveSelection from './moveSelection';
+import moveSelectionBy from './moveSelectionBy';
+import setColumnAlign from './setColumnAlign';
+
+export {
+    insertTable,
+    insertRow,
+    removeRow,
+    insertColumn,
+    removeColumn,
+    removeTable,
+    moveSelection,
+    moveSelectionBy,
+    setColumnAlign
+};
