+++ lib/changes/insertColumn.js
@@ -1,9 +1,10 @@
 // @flow
 import { type Change } from 'slate';
 import { List } from 'immutable';
-import TablePosition from '../TablePosition';
-import moveSelection from './moveSelection';
-import createCell from '../createCell';
+
+import { TablePosition, createCell } from '../utils';
+import { moveSelection } from '../changes';
+
 import ALIGN from '../ALIGN';
 
 import type Options from '../options';
