+++ lib/changes/insertRow.js
@@ -1,8 +1,7 @@
 // @flow
 import { type Change } from 'slate';
-import createRow from '../createRow';
-import TablePosition from '../TablePosition';
 
+import { TablePosition, createRow } from '../utils';
 import type Options from '../options';
 
 /**
