+++ lib/changes/insertTable.js
@@ -1,7 +1,7 @@
 // @flow
 import { type Change } from 'slate';
-import createTable from '../createTable';
 
+import { createTable } from '../utils';
 import type Options from '../options';
 
 /**
