+++ lib/changes/moveSelectionBy.js
@@ -1,8 +1,8 @@
 // @flow
 import { type Change } from 'slate';
-import TablePosition from '../TablePosition';
-import moveSelection from './moveSelection';
 
+import { TablePosition } from '../utils';
+import { moveSelection } from '../changes';
 import type Options from '../options';
 
 /**
