+++ lib/changes/removeColumn.js
@@ -2,7 +2,7 @@
 import { type Change } from 'slate';
 import { List } from 'immutable';
 
-import TablePosition from '../TablePosition';
+import { TablePosition } from '../utils';
 import type Options from '../options';
 
 /**
