+++ lib/changes/removeRow.js
@@ -1,7 +1,7 @@
 // @flow
 import { type Change } from 'slate';
 
-import TablePosition from '../TablePosition';
+import { TablePosition } from '../utils';
 import type Options from '../options';
 
 /**
