+++ lib/changes/setColumnAlign.js
@@ -1,9 +1,8 @@
 // @flow
 import { type Change } from 'slate';
 
-import TablePosition from '../TablePosition';
+import { TablePosition, createAlign } from '../utils';
 import ALIGN from '../ALIGN';
-import createAlign from '../createAlign';
 import type Options from '../options';
 
 /**
