+++ lib/core.js
@@ -1,19 +1,17 @@
 // @flow
-import insertTable from './changes/insertTable';
-import insertRow from './changes/insertRow';
-import removeRow from './changes/removeRow';
-import insertColumn from './changes/insertColumn';
-import removeColumn from './changes/removeColumn';
-import removeTable from './changes/removeTable';
-import moveSelection from './changes/moveSelection';
-import moveSelectionBy from './changes/moveSelectionBy';
-import setColumnAlign from './changes/setColumnAlign';
-
-import isSelectionInTable from './utils/isSelectionInTable';
-import getPosition from './utils/getPosition';
-
-import schema from './schema/schema';
-import validateNode from './schema/validateNode';
+import {
+    insertTable,
+    insertRow,
+    removeRow,
+    insertColumn,
+    removeColumn,
+    removeTable,
+    moveSelection,
+    moveSelectionBy,
+    setColumnAlign
+} from './changes';
+import { isSelectionInTable, getPosition } from './utils';
+import { schema, validateNode } from './validation';
 
 import ALIGN from './ALIGN';
 import Options, { type OptionsFormat } from './options';
