+++ lib/handlers/index.js
@@ -1,0 +1,8 @@
+import onEnter from './onEnter';
+import onModEnter from './onModEnter';
+import onTab from './onTab';
+import onBackspace from './onBackspace';
+import onUpDown from './onUpDown';
+import onKeyDown from './onKeyDown';
+
+export { onEnter, onModEnter, onTab, onBackspace, onUpDown, onKeyDown };
