+++ lib/handlers/onBackspace.js
@@ -1,5 +1,6 @@
 // @flow
 import { Range, type Change } from 'slate';
+
 import type Options from '../options';
 
 function onBackspace(
