+++ lib/handlers/onEnter.js
@@ -1,7 +1,8 @@
 // @flow
 import { type Change } from 'slate';
+
 import type Options from '../options';
-import insertRow from '../changes/insertRow';
+import { insertRow } from '../changes';
 
 /**
  * Insert a new row when pressing "Enter"
