+++ lib/handlers/onKeyDown.js
@@ -2,15 +2,14 @@
 
 import { type Change } from 'slate';
 
-import isSelectionInTable from './utils/isSelectionInTable';
+import type Options from '../options';
+import { isSelectionInTable } from '../utils';
 
-import onEnter from './handlers/onEnter';
-import onModEnter from './handlers/onModEnter';
-import onTab from './handlers/onTab';
-import onBackspace from './handlers/onBackspace';
-import onUpDown from './handlers/onUpDown';
-
-import type Options from './options';
+import onEnter from './onEnter';
+import onModEnter from './onModEnter';
+import onTab from './onTab';
+import onBackspace from './onBackspace';
+import onUpDown from './onUpDown';
 
 const KEY_ENTER = 'Enter';
 const KEY_TAB = 'Tab';
