+++ lib/handlers/onModEnter.js
@@ -1,7 +1,7 @@
 // @flow
 import { Block, Text, type Change } from 'slate';
 
-import TablePosition from '../TablePosition';
+import { TablePosition } from '../utils';
 import type Options from '../options';
 
 /**
