+++ lib/handlers/onTab.js
@@ -1,9 +1,8 @@
 // @flow
 import { type Change } from 'slate';
 
-import TablePosition from '../TablePosition';
-import moveSelectionBy from '../changes/moveSelectionBy';
-import insertRow from '../changes/insertRow';
+import { TablePosition } from '../utils';
+import { moveSelectionBy, insertRow } from '../changes';
 import type Options from '../options';
 
 /**
