+++ lib/handlers/onUpDown.js
@@ -1,8 +1,8 @@
 // @flow
 import { type Change } from 'slate';
 
-import TablePosition from '../TablePosition';
-import moveSelectionBy from '../changes/moveSelectionBy';
+import { TablePosition } from '../utils';
+import { moveSelectionBy } from '../changes';
 import type Options from '../options';
 
 function onUpDown(
