+++ lib/index.js
@@ -3,7 +3,7 @@
 import Options, { type OptionsFormat } from './options';
 import ALIGN from './ALIGN';
 import core from './core';
-import onKeyDown from './onKeyDown';
+import { onKeyDown } from './handlers';
 
 /**
  *  Returns the full plugin object (behavior + rendering + schema)
