+++ lib/utils/createAlign.js
@@ -1,12 +1,9 @@
 // @flow
 import { Range } from 'immutable';
-import ALIGN from './ALIGN';
+import ALIGN from '../ALIGN';
 
 /**
  * Create a set of alignment
- * @param  {Number} columns
- * @param  {List} base
- * @return {List} list
  */
 function createAlign(columns: number, base: string[] = []): string[] {
     return Range(0, columns)
