+++ lib/utils/createRow.js
@@ -1,9 +1,9 @@
 // @flow
 import { Range } from 'immutable';
 import { Block } from 'slate';
-import createCell from './createCell';
 
-import type Options from './options';
+import type Options from '../options';
+import createCell from './createCell';
 
 /**
  * Create a new row block
