+++ lib/utils/createTable.js
@@ -1,10 +1,11 @@
 // @flow
 import { Range } from 'immutable';
 import { Block } from 'slate';
+
+import type Options from '../options';
+
 import createRow from './createRow';
 import createAlign from './createAlign';
-
-import type Options from './options';
 
 /**
  * Create a table
