+++ lib/utils/getPosition.js
@@ -1,9 +1,10 @@
 // @flow
 import { type Value } from 'slate';
 
-import TablePosition from '../TablePosition';
-import isSelectionInTable from './isSelectionInTable';
 import type Options from '../options';
+
+import TablePosition from './TablePosition';
+import isSelectionInTable from './isSelectionInTable';
 
 /**
  * The position of the selection start block, in the current table
