+++ lib/utils/index.js
@@ -1,0 +1,18 @@
+import getPosition from './getPosition';
+import isSelectionInTable from './isSelectionInTable';
+import TablePosition from './TablePosition';
+
+import createAlign from './createAlign';
+import createCell from './createCell';
+import createRow from './createRow';
+import createTable from './createTable';
+
+export {
+    getPosition,
+    isSelectionInTable,
+    TablePosition,
+    createAlign,
+    createCell,
+    createRow,
+    createTable
+};
