+++ lib/validation/index.js
@@ -1,0 +1,4 @@
+import schema from './schema';
+import validateNode from './validateNode';
+
+export { schema, validateNode };
