+++ lib/validation/validateNode.js
@@ -2,7 +2,7 @@
 import { Block, Text, type Change, type Node } from 'slate';
 import { Range, List } from 'immutable';
 
-import createAlign from '../createAlign';
+import { createAlign } from '../utils';
 import type Options from '../options';
 
 type Normalizer = Change => any;
